+++ doc/Changelog.md
@@ -7,6 +7,7 @@
 * albums and media can be protected with passwords: protected content isn't exposed in the album tree or the cache
   * new option: `passwords_file`: a file which is read when the options file is read, it defines the passwords and their respective identifiers
   * new option: `passwords_marker`: the name of the files to put in the albums tree in order to specify the passwords for albums and files
+    * directory and file names can be matched case sensitive/insentitive, whole name/part of it
   * new option: `passwords_subdir`: the cache subdir where the passwords files are written
   * new shortcut `u` to unveil protected content
   * new option `protected_directories_prefix`: the prefix for the protected json files directories